Billing Component/Styling refactor
